Fix background script errors by moving constants to separate file (#834)
