+++ apps/browser-extension/src/entrypoints/background.ts
@@ -1,19 +1,15 @@
-import { browser } from "#imports";
 import { defineBackground } from '#imports';
 import { onMessage } from "webext-bridge/background";
 import { setupContextMenus } from '@/entrypoints/background/ContextMenu';
 import { handleCheckAuthStatus, handleClearVault, handleCreateIdentity, handleGetCredentials, handleGetDefaultEmailDomain, handleGetDefaultIdentityLanguage, handleGetDerivedKey, handleGetPasswordSettings, handleGetVault, handleStoreVault, handleSyncVault } from '@/entrypoints/background/VaultMessageHandler';
-import { handleOpenPopup, handlePopupWithCredential } from '@/entrypoints/background/PopupMessageHandler';
-
-type ToggleContextMenuData = {
-  enabled: boolean;
-};
-
+import { handleOpenPopup, handlePopupWithCredential, handleToggleContextMenu } from '@/entrypoints/background/PopupMessageHandler';
+import { storage } from '#imports';
+import { GLOBAL_CONTEXT_MENU_ENABLED_KEY } from '@/utils/Constants';
 export default defineBackground({
   /**
    * This is the main entry point for the background script.
    */
-  main() {
+  async main() {
     // Listen for messages using webext-bridge
     onMessage('CHECK_AUTH_STATUS', () => handleCheckAuthStatus());
     onMessage('STORE_VAULT', ({ data }) => handleStoreVault(data));
@@ -28,12 +24,12 @@ export default defineBackground({
     onMessage('GET_DERIVED_KEY', () => handleGetDerivedKey());
     onMessage('OPEN_POPUP', () => handleOpenPopup());
     onMessage('OPEN_POPUP_WITH_CREDENTIAL', ({ data }) => handlePopupWithCredential(data));
-    onMessage('TOGGLE_CONTEXT_MENU', ({ data }: { data: ToggleContextMenuData }) => {
-      if (data.enabled) {
-        setupContextMenus();
-      } else {
-        browser.contextMenus.removeAll();
-      }
-    });
+    onMessage('TOGGLE_CONTEXT_MENU', ({ data }) => handleToggleContextMenu(data));
+
+    // Setup context menus
+    const isContextMenuEnabled = await storage.getItem(GLOBAL_CONTEXT_MENU_ENABLED_KEY) ?? true;
+    if (isContextMenuEnabled) {
+      setupContextMenus();
+    }
   }
 });
