+++ apps/browser-extension/src/entrypoints/background/PopupMessageHandler.ts
@@ -1,5 +1,7 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
 import { browser } from '#imports';
 import { BoolResponse } from '@/utils/types/messaging/BoolResponse';
+import { setupContextMenus } from './ContextMenu';
 
 /**
  * Handle opening the popup.
@@ -20,7 +22,7 @@ export function handleOpenPopup() : Promise<BoolResponse> {
 /**
  * Handle opening the popup with a credential.
  */
-export function handlePopupWithCredential(message: { credentialId: string }) : Promise<BoolResponse> {
+export function handlePopupWithCredential(message: any) : Promise<BoolResponse> {
   return (async () : Promise<BoolResponse> => {
     browser.windows.create({
       url: browser.runtime.getURL(`/popup.html?expanded=true#/credentials/${message.credentialId}`),
@@ -29,6 +31,20 @@ export function handlePopupWithCredential(message: { credentialId: string }) : P
       height: 600,
       focused: true
     });
+    return { success: true };
+  })();
+}
+
+/**
+ * Handle toggling the context menu.
+ */
+export function handleToggleContextMenu(message: any) : Promise<BoolResponse> {
+  return (async () : Promise<BoolResponse> => {
+    if (!message.enabled) {
+      browser.contextMenus.removeAll();
+    } else {
+      setupContextMenus();
+    }
     return { success: true };
   })();
 }
