+++ apps/browser-extension/src/entrypoints/popup/context/AuthContext.tsx
@@ -2,7 +2,7 @@ import React, { createContext, useContext, useState, useEffect, useMemo, useCall
 import { storage } from '#imports';
 import { sendMessage } from 'webext-bridge/popup';
 import { useDb } from '@/entrypoints/popup/context/DbContext';
-import { VAULT_LOCKED_DISMISS_UNTIL_KEY } from '@/entrypoints/contentScript/Popup';
+import { VAULT_LOCKED_DISMISS_UNTIL_KEY } from '@/utils/Constants';
 
 type AuthContextType = {
   isLoggedIn: boolean;
