+++ apps/browser-extension/src/entrypoints/popup/pages/AuthSettings.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import { AppInfo } from '@/utils/AppInfo';
 import { storage } from '#imports';
-import { GLOBAL_AUTOFILL_POPUP_ENABLED_KEY, DISABLED_SITES_KEY, VAULT_LOCKED_DISMISS_UNTIL_KEY } from '@/entrypoints/contentScript/Popup';
+import { GLOBAL_AUTOFILL_POPUP_ENABLED_KEY, DISABLED_SITES_KEY, VAULT_LOCKED_DISMISS_UNTIL_KEY } from '@/utils/Constants';
 
 type ApiOption = {
   label: string;
