+++ apps/browser-extension/src/entrypoints/popup/pages/Settings.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState, useCallback } from 'react';
 import { storage } from "#imports";
 import { sendMessage } from 'webext-bridge/popup';
-import { DISABLED_SITES_KEY, GLOBAL_AUTOFILL_POPUP_ENABLED_KEY, GLOBAL_CONTEXT_MENU_ENABLED_KEY, TEMPORARY_DISABLED_SITES_KEY } from '@/entrypoints/contentScript/Popup';
+import { DISABLED_SITES_KEY, GLOBAL_AUTOFILL_POPUP_ENABLED_KEY, GLOBAL_CONTEXT_MENU_ENABLED_KEY, TEMPORARY_DISABLED_SITES_KEY } from '@/utils/Constants';
 import { AppInfo } from '@/utils/AppInfo';
 import { useTheme } from '@/entrypoints/popup/context/ThemeContext';
 import { browser } from "#imports";
