+++ apps/browser-extension/src/utils/Constants.ts
@@ -1,0 +1,10 @@
+// TODO: store generic setting constants somewhere else.
+export const DISABLED_SITES_KEY = 'local:aliasvault_disabled_sites';
+export const GLOBAL_AUTOFILL_POPUP_ENABLED_KEY = 'local:aliasvault_global_autofill_popup_enabled';
+export const GLOBAL_CONTEXT_MENU_ENABLED_KEY = 'local:aliasvault_global_context_menu_enabled';
+export const VAULT_LOCKED_DISMISS_UNTIL_KEY = 'local:aliasvault_vault_locked_dismiss_until';
+export const TEMPORARY_DISABLED_SITES_KEY = 'local:aliasvault_temporary_disabled_sites';
+
+// TODO: store these settings in the actual vault when updating the datamodel for roadmap v1.0.
+export const LAST_CUSTOM_EMAIL_KEY = 'local:aliasvault_last_custom_email';
+export const LAST_CUSTOM_USERNAME_KEY = 'local:aliasvault_last_custom_username';
